applying singuler pattern on echo and glasswall structs to read their section in config file once
